+++ ballast-idea-plugin/src/main/kotlin/com/copperleaf/ballast/debugger/idea/BallastIdeaPlugin.kt
@@ -1,3 +1,0 @@
-package com.copperleaf.ballast.debugger.idea
-
-class BallastIdeaPlugin